+++ pkg/BUILD.bazel
@@ -333,6 +333,7 @@ ALL_TESTS = [
     "//pkg/scheduledjobs/schedulebase:schedulebase_test",
     "//pkg/scheduledjobs:scheduledjobs_test",
     "//pkg/security/certmgr:certmgr_test",
+    "//pkg/security/clientcert:clientcert_test",
     "//pkg/security/password:password_test",
     "//pkg/security/sessionrevival:sessionrevival_test",
     "//pkg/security/username:username_disallowed_imports_test",
@@ -1702,6 +1703,8 @@ GO_TARGETS = [
     "//pkg/security/certmgr:certmgr",
     "//pkg/security/certmgr:certmgr_test",
     "//pkg/security/certnames:certnames",
+    "//pkg/security/clientcert:clientcert",
+    "//pkg/security/clientcert:clientcert_test",
     "//pkg/security/clientsecopts:clientsecopts",
     "//pkg/security/distinguishedname:distinguishedname",
     "//pkg/security/password:password",
