+++ pkg/security/clientcert/cert_expiry_cache.go
@@ -3,7 +3,7 @@
 // Use of this software is governed by the CockroachDB Software License
 // included in the /LICENSE file.
 
-package security
+package clientcert
 
 import (
 	"context"
